requirements.txt
Flask
beautifulsoup4
Flask-Caching
playwright
python-dotenv
flask-cors
openai
astro
express 
body-parser 
bcrypt 
jsonwebtoken
Flask-Session
PyJWT
requests
stripe
schedule
aiohttp